Move button into separate file
